Handle access to entities

+++ src/server/application/use-cases/groups/remove-group.use-case.ts
@@ -1,13 +1,26 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { IGroupsRepository } from '@/server/application/repositories/groups.repository.interface';
+import { auth } from '@/auth';
 import { IGroupMembersRepository } from '@/server/application/repositories/groupMembers.repository.interface';
+import { IGroupsRepository } from '@/server/application/repositories/groups.repository.interface';
+import { IGroupService } from '../../services/group.service.interface';
 
 export const removeGroupUseCase =
   (
     groupsRepository: IGroupsRepository,
-    groupMembersRepository: IGroupMembersRepository
+    groupMembersRepository: IGroupMembersRepository,
+    groupService: IGroupService
   ) =>
   async (groupId: string, tx?: any): Promise<boolean> => {
+    const user = (await auth())?.user;
+    if (!user) {
+      throw new Error('User not found');
+    }
+
+    await groupService.verifyUserInGroup(user.id, groupId);
+    if (!(await groupService.canUserModifyGroup(user.id, groupId))) {
+      throw new Error('User does not have permission to remove this group');
+    }
+
     // Remove all users from the group
     const membersRemoved = await groupMembersRepository.removeAllUsersFromGroup(
       groupId,
